responsibleai: Move tests into respective directories
