+++ responsibleai/tests/model_analysis/test_model_analysis.py
@@ -17,18 +17,18 @@ from responsibleai._internal.constants import ManagerNames
 from responsibleai._tools.shared.state_directory_management import \
     DirectoryManager
 
-from .causal_manager_validator import validate_causal
-from .common_utils import (create_adult_income_dataset,
+from tests.causal_manager_validator import validate_causal
+from tests.common_utils import (create_adult_income_dataset,
                            create_binary_classification_dataset,
                            create_cancer_data,
                            create_complex_classification_pipeline,
                            create_housing_data, create_iris_data,
                            create_models_classification,
                            create_models_regression)
-from .counterfactual_manager_validator import validate_counterfactual
-from .error_analysis_validator import (setup_error_analysis,
+from tests.counterfactual_manager_validator import validate_counterfactual
+from tests.error_analysis_validator import (setup_error_analysis,
                                        validate_error_analysis)
-from .explainer_manager_validator import setup_explainer, validate_explainer
+from tests.explainer_manager_validator import setup_explainer, validate_explainer
 
 LABELS = 'labels'
