+++ responsibleai/tests/rai_insights/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) Microsoft Corporation
+# Licensed under the MIT License.
+
+"""RAIInsights tests."""
